Don't try to open() folders.  This fixes a build issue on the bots which use svn.  This doesn't effect git users which is why it didn't catch it earlier.
BUG=

Review URL: https://codereview.chromium.org/11099027

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@1437 050acbb0-2703-11df-ab0a-9f3f633ae91d

+++ visual_studio/NativeClientVSAddIn/create_package.py
@@ -148,8 +148,12 @@ def main():
     shutil.rmtree(dest32)
   os.makedirs(dest32)
   for filename in os.listdir(src64):
+    srcfile = join(src64, filename)
+    # skip non-files, in particular .svn folders.
+    if not os.path.isfile(srcfile):
+      continue
     destfile = join(dest32, filename.replace('NaCl64', 'NaCl32'))
-    data = open(join(src64, filename), "rb").read()
+    data = open(srcfile, "rb").read()
     data = data.replace("x86_64", "i686")
     data = data.replace("64", "32")
     open(destfile, "wb").write(data)
